added tests for fields

+++ src/FarmAdvisor.Models/SensorModels/SensorUpdateModel.cs
@@ -5,7 +5,7 @@ namespace FarmAdvisor.Models
     {
         public Guid? FieldId { get; set; }
         public string? SerialNo { get; set; }
-        public int? GDD { get; set; }
+        public double? GDD { get; set; }
         public int? DefaultGDD { get; set; }
         public double? Long { get; set; }
         public double? Lat { get; set; }
